backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
DateTime==5.5
fastapi==0.119.0
google-api-core>=1.34.1,<3.0.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos>=1.50.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
lxml==6.0.2
oauth2client==4.1.3
proto-plus==1.26.1
protobuf>=3.19.5,<5.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic_core==2.41.4
email-validator==2.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
zope.interface==8.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography>=41.0.0

# Database
sqlalchemy==2.0.36
python-dotenv==1.0.0

# Supabase
supabase==2.23.3

# AI & LangChain
langchain==0.2.16
langchain-google-genai==1.0.10
google-generativeai==0.7.2

# ML Classification
scikit-learn==1.6.1
sentence-transformers>=2.7.0
